requirements.txt
astroid==2.3.3
Click==7.0
colorama==0.4.3
Flask==1.1.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
pylint==2.4.4
PyMySQL==0.9.3
PyQt5==5.14.1
PyQt5-sip==12.7.1
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==0.16.1
wrapt==1.11.2
